Revert "feat: add page indexes for metadata (#1027)"
This reverts commit 1fd7857ed1f092cdf21f5fa5007fe0391eb54345.
