+++ analytic_engine/src/sst/reader.rs
@@ -15,17 +15,20 @@ pub mod error {
     #[derive(Debug, Snafu)]
     #[snafu(visibility(pub))]
     pub enum Error {
-        #[snafu(display("Try to read again, path:{path}.\nBacktrace:\n{backtrace}"))]
+        #[snafu(display("Try to read again, path:{}.\nBacktrace:\n{}", path, backtrace))]
         ReadAgain { backtrace: Backtrace, path: String },
 
-        #[snafu(display("Fail to read persisted file, path:{path}, err:{source}"))]
+        #[snafu(display("Fail to read persisted file, path:{}, err:{}", path, source))]
         ReadPersist { path: String, source: GenericError },
 
-        #[snafu(display("Failed to decode record batch, err:{source}"))]
+        #[snafu(display("Failed to decode record batch, err:{}", source))]
         DecodeRecordBatch { source: GenericError },
 
         #[snafu(display(
-            "Failed to decode sst meta data, file_path:{file_path}, err:{source}.\nBacktrace:\n{backtrace:?}",
+            "Failed to decode sst meta data, file_path:{}, err:{}.\nBacktrace:\n{:?}",
+            file_path,
+            source,
+            backtrace
         ))]
         FetchAndDecodeSstMeta {
             file_path: String,
@@ -33,52 +36,43 @@ pub mod error {
             backtrace: Backtrace,
         },
 
-        #[snafu(display(
-            "Failed to decode page indexes for meta data, file_path:{file_path}, err:{source}.\nBacktrace:\n{backtrace:?}",
-        ))]
-        DecodePageIndexes {
-            file_path: String,
-            source: parquet::errors::ParquetError,
-            backtrace: Backtrace,
-        },
-
-        #[snafu(display("Failed to decode sst meta data, err:{source}"))]
+        #[snafu(display("Failed to decode sst meta data, err:{}", source))]
         DecodeSstMeta { source: GenericError },
 
-        #[snafu(display("Sst meta data is not found.\nBacktrace:\n{backtrace}"))]
+        #[snafu(display("Sst meta data is not found.\nBacktrace:\n{}", backtrace))]
         SstMetaNotFound { backtrace: Backtrace },
 
-        #[snafu(display("Fail to projection, err:{source}"))]
+        #[snafu(display("Fail to projection, err:{}", source))]
         Projection { source: GenericError },
 
-        #[snafu(display("Sst meta data is empty.\nBacktrace:\n{backtrace}"))]
+        #[snafu(display("Sst meta data is empty.\nBacktrace:\n{}", backtrace))]
         EmptySstMeta { backtrace: Backtrace },
 
-        #[snafu(display("Invalid schema, err:{source}"))]
+        #[snafu(display("Invalid schema, err:{}", source))]
         InvalidSchema { source: common_types::schema::Error },
 
-        #[snafu(display("Meet a datafusion error, err:{source}\nBacktrace:\n{backtrace}"))]
+        #[snafu(display("Meet a datafusion error, err:{}\nBacktrace:\n{}", source, backtrace))]
         DataFusionError {
             source: datafusion::error::DataFusionError,
             backtrace: Backtrace,
         },
 
-        #[snafu(display("Meet a object store error, err:{source}\nBacktrace:\n{backtrace}"))]
+        #[snafu(display("Meet a object store error, err:{}\nBacktrace:\n{}", source, backtrace))]
         ObjectStoreError {
             source: object_store::ObjectStoreError,
             backtrace: Backtrace,
         },
 
-        #[snafu(display("Meet a parquet error, err:{source}\nBacktrace:\n{backtrace}"))]
+        #[snafu(display("Meet a parquet error, err:{}\nBacktrace:\n{}", source, backtrace))]
         ParquetError {
             source: parquet::errors::ParquetError,
             backtrace: Backtrace,
         },
 
-        #[snafu(display("Other kind of error:{source}"))]
+        #[snafu(display("Other kind of error:{}", source))]
         Other { source: GenericError },
 
-        #[snafu(display("Other kind of error, msg:{msg}.\nBacktrace:\n{backtrace}"))]
+        #[snafu(display("Other kind of error, msg:{}.\nBacktrace:\n{}", msg, backtrace))]
         OtherNoCause { msg: String, backtrace: Backtrace },
     }
