+++ components/parquet_ext/Cargo.toml
@@ -17,8 +17,6 @@ async-trait = { workspace = true }
 bytes = { workspace = true }
 common_util = { workspace = true }
 datafusion = { workspace = true }
-futures = { workspace = true }
 log = { workspace = true }
-object_store = { workspace = true }
 parquet = { workspace = true }
 tokio = { workspace = true }
