+++ components/parquet_ext/src/lib.rs
@@ -2,7 +2,6 @@
 
 pub mod meta_data;
 pub mod prune;
-pub mod reader;
 pub mod reverse_reader;
 #[cfg(test)]
 pub mod tests;
